Add Phiro tests

+++ catroid/src/androidTest/java/org/catrobat/catroid/test/content/actions/SetColorActionTest.java
@@ -45,7 +45,7 @@ public class SetColorActionTest extends InstrumentationTestCase {
 	}
 
 	public void testColorEffect() {
-		assertEquals("Unexpected initial color value", (int) 0, (int) sprite.look.getColorInUserInterfaceDimensionUnit());
+		assertEquals("Unexpected initial color value", 0, (int) sprite.look.getColorInUserInterfaceDimensionUnit());
 		sprite.getActionFactory().createSetColorAction(sprite, color).act(1.0f);
 		assertEquals("Incorrect color value after SetColorTo executed", COLOR, sprite.look.getColorInUserInterfaceDimensionUnit());
 		sprite.getActionFactory().createSetColorAction(sprite, color);
@@ -54,7 +54,7 @@ public class SetColorActionTest extends InstrumentationTestCase {
 	public void testValueAboveMax() {
 		final float highColor = 1000;
 
-		assertEquals("Unexpected initial color value", (int) 0, (int) sprite.look.getColorInUserInterfaceDimensionUnit());
+		assertEquals("Unexpected initial color value", 0, (int) sprite.look.getColorInUserInterfaceDimensionUnit());
 		sprite.getActionFactory().createSetColorAction(sprite, new Formula(highColor)).act(1.0f);
 		assertEquals("Incorrect color value after SetColorTo executed", (highColor % 200), sprite.look.getColorInUserInterfaceDimensionUnit());
 	}
